GET, POST:
This commit implements the 'GET' api that is being used in the client to access data for the 'My-Cart' page. This commit also implements the 'POST' api, that is being used for the 'Add Product' page. 'Add Product' page reads the input from the user and calls the POST api in the backend to upload the data into the MongoDB's database, named showsCollection.