PUT, POST: for inserting and updating:
This commit implements the PUT api for inserting and updating the products that the user added to their cart. Inside the PUT api, it first gets the user id, then it checks if the user already axists in the database or not. If the user exists then it simply updates the user's cart, otherwise it inserts the new user along with their cart information.